# Improve performance of `last_value` by implementing special `GroupsAccumulator`